SharedPreference에 삭제한 이미지 갯수, 용량 저장하는 로직 및 화면 구현

+++ model/src/main/java/com/anliban/team/hippho/model/Image.kt
@@ -14,5 +14,5 @@ data class Image(
     val fileName: String,
     val date: Date,
     val contentUri: String,
-    val fileSize: Double
+    val fileSize: Long
 ) : Parcelable
